+++ up42/base.py
@@ -6,8 +6,8 @@ from typing import Any, Optional, Union
 import pystac_client
 import requests
 
-from up42 import auth as up42_auth
 from up42 import host, utils
+from up42.http import client, oauth
 
 logger = utils.get_logger(__name__, level=logging.INFO)
 
@@ -25,16 +25,21 @@ def _authenticated(value: Any):
 
 
 class _Workspace:
-    _auth: Optional[up42_auth.Auth] = None
     _id: Optional[str] = None
-
-    @property
-    def auth(self):
-        return _authenticated(self._auth)
+    _session: Optional[requests.Session] = None
+    _auth: Optional[oauth.Up42Auth] = None
 
     @property
     def id(self):
         return _authenticated(self._id)
+
+    @property
+    def session(self):
+        return _authenticated(self._session)
+
+    @property
+    def auth(self):
+        return _authenticated(self._auth)
 
     def authenticate(
         self,
@@ -51,13 +56,13 @@ class _Workspace:
             username: The username for the UP42 account (email UP42 console).
             password: Password for the UP42 console login.
         """
-        self._auth = up42_auth.Auth(
-            cfg_file=cfg_file,
-            username=username,
-            password=password,
-        )
+        credential_sources = client.collect_credentials(cfg_file, username, password)
+        up42_client = client.create(credential_sources, host.token_endpoint())
+        logger.info("Authentication with UP42 successful!")
         url = host.endpoint("/users/me")
-        self._id = self.auth.session.get(url).json()["data"]["id"]
+        self._session = up42_client.session
+        self._id = self.session.get(url).json()["data"]["id"]
+        self._auth = up42_client.auth
 
     def get_credits_balance(self) -> dict:
         """
@@ -67,7 +72,7 @@ class _Workspace:
             A dict with the balance of credits available in your account.
         """
         url = host.endpoint("/v2/payments/balances")
-        balance = self.auth.session.get(url).json()["available"]["amount"]
+        balance = self.session.get(url).json()["available"]["amount"]
         return {"balance": balance}
 
 
@@ -79,7 +84,7 @@ get_credits_balance = workspace.get_credits_balance
 
 class Session:
     def __get__(self, obj, obj_type=None) -> requests.Session:
-        return workspace.auth.session
+        return workspace.session
 
 
 class WorkspaceId:
@@ -96,4 +101,4 @@ class WorkspaceId:
 
 class StacClient:
     def __get__(self, obj, obj_type=None) -> pystac_client.Client:
-        return utils.stac_client(workspace.auth.client.auth)
+        return utils.stac_client(workspace.auth)