+++ CHANGELOG.md
@@ -29,6 +29,15 @@ You can check your current version with the following command:
     ```
 
 For more information, see [UP42 Python package description](https://pypi.org/project/up42-py/).
+## 2.1.1a10
+
+**Dec 2, 2024**
+- Remove duplication of workspace mocking in tests.
+- Remove duplication of setting raising session in tests.
+- Move `collection_credentials` from `auth.py` to `client.py`.
+- Dissolve `auth.Auth` in `_Workspace::authenticate`.
+- Drop legacy `auth.py`.
+
 ## 2.1.1a9
 
 **Nov 29, 2024**